+++ apps/test-suite/TestModules.js
@@ -95,6 +95,7 @@ export function getTestModules() {
       require('./tests/Contacts'),
       // require('./tests/SVG'),
       require('./tests/Localization'),
+      require('./tests/Recording'),
       optionalRequire(() => require('./tests/NewNotifications')),
       LocationTestScreen
     );