+++ apps/test-suite/tests/Recording.js
@@ -255,7 +255,12 @@ export async function test(t) {
         await recordingObject.prepareToRecordAsync(Audio.RECORDING_OPTIONS_PRESET_LOW_QUALITY);
         await recordingObject.startAsync();
         await waitFor(defaultRecordingDurationMillis);
-        t.expect(recordingObject.getURI()).toContain('file:///');
+        if (Platform.OS === 'web') {
+          // On web, URI is not available until completion
+          t.expect(recordingObject.getURI()).toEqual(null);
+        } else {
+          t.expect(recordingObject.getURI()).toContain('file:///');
+        }
         await recordingObject.stopAndUnloadAsync();
       });
     });
@@ -329,8 +334,11 @@ export async function test(t) {
               const { sound } = await recordingObject.createNewLoadedSound();
               await retryForStatus(sound, { isBuffering: false });
               const status = await sound.getStatusAsync();
-              // Android is slow and we have to take it into account when checking recording duration.
-              t.expect(status.durationMillis).toBeGreaterThan(recordingDuration * (7 / 10));
+              // Web doesn't return durations in Chrome - https://bugs.chromium.org/p/chromium/issues/detail?id=642012
+              if (Platform.OS !== 'web') {
+                // Android is slow and we have to take it into account when checking recording duration.
+                t.expect(status.durationMillis).toBeGreaterThan(recordingDuration * (7 / 10));
+              }
               t.expect(sound).toBeDefined();
             } catch (err) {
               error = err;
@@ -418,8 +426,12 @@ export async function test(t) {
               const { sound } = await recordingObject.createNewLoadedSoundAsync();
               await retryForStatus(sound, { isBuffering: false });
               const status = await sound.getStatusAsync();
-              // Android is slow and we have to take it into account when checking recording duration.
-              t.expect(status.durationMillis).toBeGreaterThan(recordingDuration * (6 / 10));
+
+              // Web doesn't return durations in Chrome - https://bugs.chromium.org/p/chromium/issues/detail?id=642012
+              if (Platform.OS !== 'web') {
+                // Android is slow and we have to take it into account when checking recording duration.
+                t.expect(status.durationMillis).toBeGreaterThan(recordingDuration * (6 / 10));
+              }
               t.expect(sound).toBeDefined();
             } catch (err) {
               error = err;